- added new controls for read-only drives/Files - added new controls when reading/writing on a vfs mounted storage - added a way to restart program in debug mode (GUI) + english mistake #56 - added a trick to mount fs on resized USER (bug in emunand creation tool is still to be fixed) - fixed the way vfs populate files with attribute 0x30 (nx archive/ directory) - fix #36 - fixed fs mounting for splitted input

+++ NxNandManager/NxHandle.h
@@ -163,8 +163,8 @@ class NxHandle {
         bool write(void *buffer, DWORD* bytesWrite, DWORD length = 0);
         bool write(u64 offset, void *buffer, DWORD* bytesWrite, DWORD length = 0);
         bool write(u32 sector, void *buffer, DWORD* bw, DWORD length);
-        bool createFile(wchar_t *path, int io_mode = GENERIC_READ);
-        void createHandle(int io_mode = GENERIC_READ);
+        bool createFile(wchar_t *path, unsigned long io_mode = GENERIC_READ);
+        void createHandle(unsigned long io_mode = GENERIC_READ);
         bool hash(u64* bytesCount);
         bool setPointer(u64 offset);
         bool dismountVolume();
